Fix bot behavior when a PR skips the merged state
The current logic doesn't account for the fact
that applying a label to a PR doesn't update
the in-memory representation of that object.

This was allowing a PR to be marked as released,
and then immediately marked as merged, in
the same run.

This did not manifest if the PR was marked as
merged in one run, but released in a later one